service1/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
gunicorn==21.2.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
urllib3==2.4.0
